Update WarfareExpanded_Text_Changes_ko_KR.xml
Hello, I made a comment yesterday for the translation error at line 30. First, sorry for less explanation for the comment that I made.
In line 30, you didn't translate two sentence at the back. So, that comment for yesterday was just for adding. I add the first sentence this time! Have a nice day!